+++ TextPointExtensibility/IPlayer.cs
@@ -15,7 +15,7 @@ namespace TextPoint
         string Timestamp();
         string Filename();
         void Speed(double speed);
-        double GetLength();
+        int GetLength();
         void PlayFrom(int sec);
         int CurrentPosition();
